fix msvc build

+++ benchmark/vkbenchncnn.cpp
@@ -74,7 +74,7 @@ public:
             cmd.wait();
 
             #pragma omp parallel for
-            for (size_t i=0; i<layers.size(); i++)
+            for (int i=0; i<layers.size(); i++)
             {
                 Layer* layer = layers[i];
